+++ lib/flutter_modular.dart
@@ -9,4 +9,8 @@ export 'src/inject/inject.dart';
 export 'src/inject/bind.dart';
 export 'src/routers/router.dart';
 export 'src/widgets/modular_app.dart';
+export 'src/widgets/module_widget.dart';
 export 'src/widgets/consumer_widget.dart';
+export 'src/widgets/modular_stateless_widget.dart';
+export 'src/widgets/modular_stateful_widget_state.dart';
+export 'src/exceptions/modular_error.dart';