+++ lib/src/exceptions/modular_error.dart
@@ -1,0 +1,10 @@
+class ModularError implements Exception {
+  final String message;
+
+  ModularError(this.message);
+
+  @override
+  String toString() {
+    return "ModularError: $message";
+  }
+}