+++ lib/src/routers/router.dart
@@ -21,11 +21,10 @@ class Router {
   }) {
     assert(routerName != null);
 
-    if(transition == null)
-      throw ArgumentError('transaction must not be null');
-    if(module == null && child == null)
-       throw ArgumentError('[module] or [child] must be provided');
-    if(module != null && child != null)
+    if (transition == null) throw ArgumentError('transition must not be null');
+    if (module == null && child == null)
+      throw ArgumentError('[module] or [child] must be provided');
+    if (module != null && child != null)
       throw ArgumentError('You should provide only [module] or [child]');
   }