+++ test/app/app_module.dart
@@ -21,8 +21,12 @@ class AppModule extends MainModule {
           guards: [MyGuard()],
           transition: TransitionType.fadeIn,
         ),
-        Router("/", module: HomeModule(), transition: TransitionType.fadeIn,),
-        Router("/home", module: HomeModule()),
+        Router(
+          "/",
+          module: HomeModule(),
+          transition: TransitionType.fadeIn,
+        ),
+        Router("/home", module: HomeModule(), guards: [MyGuard()]),
       ];
 
   @override