+++ test/app/modules/home/home_bloc.dart
@@ -1,12 +1,15 @@
 import 'package:flutter_modular/flutter_modular.dart';
 
-class HomeBloc extends Disposable {
+import '../../app_bloc.dart';
 
+class HomeBloc extends Disposable {
   String testingText = 'testing inject';
+  final AppBloc app;
+
+  HomeBloc(this.app);
 
   @override
   void dispose() {
     print('call dispose');
   }
-
 }